Add files via upload
There seems to be a corellation between ['waitng', 'eruption] points of data. The data suggests there is a cyclic activity that happens below Yellowstone that allows for a short waiting time and a short eruption followed by a longer waiting time with a longer eruption that could especially be seen on the dist and scatter plots. If the dataset is equipped with the date, time and also the height of the eruption, a more precise conclusion could be made to either support or rebut the hypothesis.